Fix warnings

+++ src/modules/backends/web/qtwebkit/QtWebKitFtpListingNetworkReply.cpp
@@ -282,7 +282,7 @@ qint64 QtWebKitFtpListingNetworkReply::readData(char *data, qint64 maxSize)
 	{
 		const qint64 number(qMin(maxSize, (m_content.size() - m_offset)));
 
-		memcpy(data, (m_content.constData() + m_offset), number);
+		memcpy(data, (m_content.constData() + m_offset), static_cast<size_t>(number));
 
 		m_offset += number;
